+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
 # Changelog
 
-#### 10th December 2024 - Version 2.0.25
+#### 10th December 2024 - Version 2.0.26
 
+-   Improvement: Added button style attributes
 -   Improvement: Force scroll behavior to auto to avoid conflict bewteen theme JS scroll features and Elementor
 
 #### 5th December 2024 - Version 2.0.24